+++ react-components/src/components/CardsList/CardsList.tsx
@@ -1,33 +1,30 @@
 import { FC } from 'react'
 
+import { Product } from '../../pages/HomePage/hooks'
 import CardHome from '../CardHome'
-import { CardsListProps } from './types'
-import { useCardsListHome } from './hooks'
 
 import styles from './CardsList.module.scss'
 
-const CardsList: FC<CardsListProps> = ({ searchValue }) => {
-  const { filterProducts, isLoaded, error } = useCardsListHome(searchValue)
+export interface CardsListProps {
+  products: Product[]
+}
 
+const CardsList: FC<CardsListProps> = ({ products }) => {
   return (
-    <>
-      {!isLoaded && <div>Loading...</div>}
-      {error && <div>Error: {error}</div>}
-      <div className={styles.wrapperCardsList}>
-        <div className={styles.cards}>
-          {filterProducts.map(({ title, price, image, description }, index) => (
-            <div key={index}>
-              <CardHome
-                title={title}
-                price={price}
-                image={image}
-                desc={description}
-              />
-            </div>
-          ))}
-        </div>
+    <div className={styles.wrapperCardsList}>
+      <div className={styles.cards}>
+        {products.map(({ title, price, image, description }, index) => (
+          <div key={index}>
+            <CardHome
+              title={title}
+              price={price}
+              image={image}
+              desc={description}
+            />
+          </div>
+        ))}
       </div>
-    </>
+    </div>
   )
 }
