--- react-components/src/components/CardsList/types.ts
@@ -1,25 +1,0 @@
-export interface Rating {
-  rate: number
-  count: number
-}
-
-export interface Product {
-  title: string
-  price: number
-  description: string
-  category: string
-  image: string
-  id?: number
-  rating?: Rating
-}
-
-export interface ResponseProducts {
-  error: null | string
-  isLoaded: boolean
-  items: Product[]
-  search: string
-}
-
-export interface CardsListProps {
-  searchValue: string
-}
