+++ react-components/src/pages/HomePage/hooks/hooks.ts
@@ -1,12 +1,12 @@
 import { useState } from 'react'
-import { useDebounce } from '../../hooks'
+import { useDebounce } from '../../../hooks'
 
 export const useHomePage = () => {
   const [searchString, setSearchString] = useState<string>('')
 
-  const searchDebounced = useDebounce(searchString, 300)
+  // const searchDebounced = useDebounce(searchString.toUpperCase(), 300)
 
-  return { searchString: searchDebounced, setSearchString }
+  return { searchString, setSearchString }
 }
 
 // const setSearchValue = (value: string) => {
