+++ react-components/src/pages/HomePage/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useFetchAllProducts'
+export * from './hooks'
